backend/requirements.txt
# ═══════════════════════════════════════════════
# INTELLISTRUCT – COMPOSITE BAR ANALYZER
# Backend Dependencies
# ═══════════════════════════════════════════════

fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic[email]>=2.0.0
motor>=3.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.6
bcrypt>=4.0.0
PyJWT>=2.8.0
pytest>=7.4.0
httpx>=0.24.0
